refactor(selection): migrate DomSelectionBridge tests and delete bridge files
